feat: auth implementation

+++ resources/views/image/index.blade.php
@@ -10,15 +10,17 @@
             @foreach($images as $image)
                 <div class="col-sm-6 col-lg-4 mb-4">
                     <div class="card">
-                        <a href="{{ $image->permalink()}}">
+                        <a href="{{ $image->permalink()}}">s
                             <img src="{{ $image->fileUrl() }}" height="100%" alt="{{ $image->title }}" class="card-img-top">
                         </a>
-                        <div class="photo-buttons">
-                            <a href="{{ $image->route('edit')}}" class="btn btn-sm btn-info me-2">Edit</a>
-                            <x-form action="{{ $image->route('destroy') }}" method="delete" method="delete">
-                                <button class="btn btn-sm btn-danger">Delete</button>
-                            </x-form>
-                        </div>
+                        @if(auth()->check() && auth()->user()->can('update', $image))
+                            <div class="photo-buttons">
+                                <a href="{{ $image->route('edit')}}" class="btn btn-sm btn-info me-2">Edit</a>
+                                <x-form action="{{ $image->route('destroy') }}" method="delete" method="delete">
+                                    <button class="btn btn-sm btn-danger">Delete</button>
+                                </x-form>
+                            </div>
+                        @endif
                     </div>
                 </div>
             @endforeach
